Create new type `FeedImageLoaderTestCase` to add the `expect` method that is going to be reused on feed loader image tests
