+++ EssentialApp/EssentialAppTests/FeedImageDataLoaderWithFallBackCompositeTests.swift
@@ -9,7 +9,7 @@ import XCTest
 import EssentialFeed
 import EssentialApp
 
-class FeedImageDataLoaderWithFallBackCompositeTests: XCTestCase {
+class FeedImageDataLoaderWithFallBackCompositeTests: XCTestCase, FeedImageLoaderTestCase {
     
     func test_load_doesNotLoadImageUponCreation() {
         let (_, primaryLoader, fallbackLoader) = makeSUT()
@@ -91,29 +91,6 @@ class FeedImageDataLoaderWithFallBackCompositeTests: XCTestCase {
     }
     
     // MARK: - Helpers
-    
-    private func expect(_ sut: FeedImageDataLoader, toCompleteWith expectedResult: (FeedImageDataLoader.Result), when action: () -> Void, file: StaticString = #file, line: UInt = #line) {
-        let exp = expectation(description: "Wait for load completion")
-
-        _ = sut.loadImageData(from: anyURL()) { receivedResult in
-            switch (receivedResult, expectedResult) {
-            case let (.success(receivedFeed), .success(expectedFeed)):
-                XCTAssertEqual(receivedFeed, expectedFeed, file: file, line: line)
-
-            case (.failure, .failure):
-                break
-
-            default:
-                XCTFail("Expected \(expectedResult), got \(receivedResult) instead", file: file, line: line)
-            }
-
-            exp.fulfill()
-        }
-
-        action()
-
-        wait(for: [exp], timeout: 1.0)
-    }
     
     private func makeSUT(file: StaticString = #file, line: UInt = #line) -> (sut: FeedImageDataLoader, primaryLoader: FeedImageLoaderSpy, fallbackLoader: FeedImageLoaderSpy) {
         let primaryLoader = FeedImageLoaderSpy()
