+++ EssentialApp/EssentialAppTests/FeedLoaderCacheImageDecoratorTests.swift
@@ -35,7 +35,7 @@ class FeedLoaderCacheImageDecorator: FeedImageDataLoader {
     }
 }
 
-class FeedLoaderCacheImageDecoratorTests: XCTestCase, FeedLoaderTestCase {
+class FeedLoaderCacheImageDecoratorTests: XCTestCase, FeedImageLoaderTestCase {
     
     func test_init_doesNotLoadImageData() {
         let (_, loader) = makeSUT()
@@ -130,28 +130,4 @@ class FeedLoaderCacheImageDecoratorTests: XCTestCase, FeedLoaderTestCase {
         }
         
     }
-    
-    private func expect(_ sut: FeedImageDataLoader, toCompleteWith expectedResult: (FeedImageDataLoader.Result), when action: () -> Void, file: StaticString = #file, line: UInt = #line) {
-        let exp = expectation(description: "Wait for load completion")
-
-        _ = sut.loadImageData(from: anyURL()) { receivedResult in
-            switch (receivedResult, expectedResult) {
-            case let (.success(receivedFeed), .success(expectedFeed)):
-                XCTAssertEqual(receivedFeed, expectedFeed, file: file, line: line)
-
-            case (.failure, .failure):
-                break
-
-            default:
-                XCTFail("Expected \(expectedResult), got \(receivedResult) instead", file: file, line: line)
-            }
-
-            exp.fulfill()
-        }
-
-        action()
-
-        wait(for: [exp], timeout: 1.0)
-    }
-    
 }
