+++ EssentialApp/EssentialAppTests/Helpers/XCTestCase+FeedImageLoader.swift
@@ -1,0 +1,36 @@
+//
+//  XCTestCase+FeedImageLoader.swift
+//  EssentialAppTests
+//
+//  Created by Jair Moreno Gaspar on 11/02/21.
+//
+
+import XCTest
+import EssentialFeed
+
+protocol FeedImageLoaderTestCase: XCTestCase {}
+
+extension FeedImageLoaderTestCase {
+    func expect(_ sut: FeedImageDataLoader, toCompleteWith expectedResult: (FeedImageDataLoader.Result), when action: () -> Void, file: StaticString = #file, line: UInt = #line) {
+        let exp = expectation(description: "Wait for load completion")
+
+        _ = sut.loadImageData(from: anyURL()) { receivedResult in
+            switch (receivedResult, expectedResult) {
+            case let (.success(receivedFeed), .success(expectedFeed)):
+                XCTAssertEqual(receivedFeed, expectedFeed, file: file, line: line)
+
+            case (.failure, .failure):
+                break
+
+            default:
+                XCTFail("Expected \(expectedResult), got \(receivedResult) instead", file: file, line: line)
+            }
+
+            exp.fulfill()
+        }
+
+        action()
+
+        wait(for: [exp], timeout: 1.0)
+    }
+}
